Refactor common activity and fix bug in Liquid handler
