+++ src/scripting/Elsa.Scripting.Liquid/Services/LiquidHandler.cs
@@ -33,7 +33,7 @@ namespace Elsa.Scripting.Liquid.Services
 
         private async Task<TemplateContext> CreateTemplateContextAsync(ActivityExecutionContext activityExecutionContext, CancellationToken cancellationToken)
         {
-            var context = new TemplateContext();
+            var context = new TemplateContext(new TemplateOptions());
             context.SetValue("ActivityExecutionContext", activityExecutionContext);
             context.Model = activityExecutionContext;
             await _mediator.Publish(new EvaluatingLiquidExpression(context, activityExecutionContext), cancellationToken);
